+++ frontend/src/App.css
@@ -14,7 +14,7 @@
 }
 
 .App-header {
-  background-color: #282c34;
+  background-color: #074edd;
   min-height: 100vh;
   display: flex;
   flex-direction: column;
